Delete and valid EVENT is Done

+++ src/app.js
@@ -9,6 +9,7 @@ import './scss/main.scss'
 import './scss/select.scss'
 
 //code
+const BASE_URL = 'https://event-team-calendar-default-rtdb.europe-west1.firebasedatabase.app/data-event'
 const arrMembers = [
     {label: 'Homer J. Simpson', value: '0'},
     {label: 'Marge Simpson', value: '1'},
@@ -41,8 +42,6 @@ let currentHour = '';
 let currentDay = '';
 
 let stateEvent = [];
-let stateLength = 0;
-
 
 const getCurrentHour = (date) => {
     let hours = date.getHours();
@@ -70,24 +69,27 @@ currentHour = getCurrentHour(new Date);
 
 //Axios
 async function makeGetRequest() {
-    const res = await axios.get('https://event-team-calendar-default-rtdb.europe-west1.firebasedatabase.app/data-event.json')
+    const res = await axios.get(`${BASE_URL}.json`)
     return stateEvent = res.data;
 }
 
 async function makePostRequest(data) {
-    await axios.post('https://event-team-calendar-default-rtdb.europe-west1.firebasedatabase.app/data-event.json', data).then((res) => {
-        return console.log(res.data)
-    }).catch((error)=> {
-        console.log('Error', error)
-    });
+    const res = axios.post(`${BASE_URL}.json`, data)
 }
 
 async function deleteEventRequest(id) {
-    const res = await axios.delete('https://event-team-calendar-default-rtdb.europe-west1.firebasedatabase.app/data-event/' + id +'.json')
+    const res = await axios.delete(`${BASE_URL}/${id}.json`)
     return console.log('axios-del', res.data)
 }
 
-function formAction(){
+//Create Event
+function formAction(data) {
+    const InfoAlert = document.getElementById('formAlert');
+    let prevState;
+    data.then(function (result) {
+        return prevState = result
+    })
+
     const instance = new SelectPure(".select", {
         options: arrMembers,
         multiple: true
@@ -116,26 +118,64 @@ function formAction(){
         let participantsValue = instance.value()
         let dayValue = instance2.value();
         let timeValue = instance3.value();
-        let newState = {id: stateLength+1, name: nameValue.value, day: dayValue, time: timeValue, member: participantsValue }
+        let newState = {name: nameValue.value, day: dayValue, time: timeValue, member: participantsValue}
 
-        console.log(stateLength)
+        if (!!newState.name || !!newState.day || !!newState.time || !!newState.member) {
+            if (Object.values(prevState).filter(el => el.day === newState.day).length > 0 && Object.values(prevState).filter(el => el.time === newState.time).length > 0){
+                InfoAlert.append('Filed to create as event. Time slot is already booking')
+                InfoAlert.classList.add('alert-danger')
+                setTimeout(() => {
+                    InfoAlert.innerHTML = ''
+                    InfoAlert.classList.remove('alert-danger');
+                }, 3500)
+            } else {
+                nameValue.value = '';
+                instance.reset();
+                instance2.reset();
+                instance3.reset();
+                makePostRequest(newState).then(() => {
+                    InfoAlert.append('The event was successfully added to the calendar.')
+                    InfoAlert.classList.add('alert-success')
+                    setTimeout(() => {
+                        InfoAlert.innerHTML = ''
+                        InfoAlert.classList.remove('alert-success');
+                        window.location.href = '/'
+                    }, 3500)
 
-        nameValue.value = '';
-        instance.reset();
-        instance2.reset();
-        instance3.reset();
-        makePostRequest(newState).then((res) => {
-            //window.location.href = '/'
-        });
+                });
+            }
+        } else {
+            InfoAlert.append('All field can not be blank!')
+            InfoAlert.classList.add('alert-warning')
+            setTimeout(() => {
+                InfoAlert.innerHTML = ''
+                InfoAlert.classList.remove('alert-warning');
+            }, 3500)
+        }
     })
+}
+
+function CreateEventCell(data){
+    const eventsCell = document.getElementsByClassName('js-cell');
+    for(let k = 0; k<eventsCell.length; k++ ){
+        Object.entries(data).map((item) => {
+            if (eventsCell[k].dataset.day === item[1].day && eventsCell[k].dataset.time === item[1].time){
+                console.log('test1')
+                console.log(data)
+                eventsCell[k].innerHTML = '';
+                eventsCell[k].setAttribute('data-member', item[1].member)
+                eventsCell[k].setAttribute('data-id', item[0])
+                eventsCell[k].append(item[1].name);
+            }
+        })
+    }
 }
 
 if (window.location.pathname !== '/'){
     window.onload = function(){
-        makeGetRequest().then(()=>{
-            stateLength = Object.keys(stateEvent).length;
-        })
-        formAction();
+        let state = makeGetRequest()
+        formAction(state);
+
     }
 } else {
     makeGetRequest().then(()=> {
@@ -149,31 +189,37 @@ if (window.location.pathname !== '/'){
             const modalMember = document.getElementById('eventModal').getElementsByClassName('js-member-in-modal')[0];
             const deleteButton = document.getElementById('deleteEvent');
             let eventId = targetCell.dataset.id
-            let memberCell = targetCell.dataset.member.split(',');
             modalMember.innerHTML = '';
-
-            deleteButton.setAttribute('data-deleteId', eventId);
-
-            Object.values(arrMembers).map((member) => {
-                for (let i = 0; i<memberCell.length; i++){
-                    if (member.value === memberCell[i]){
-                        const itemMember = document.createElement('li');
-                        modalMember.append(itemMember)
-                        itemMember.append(`${member.label}`);
+            if (targetCell.dataset.member){
+                let memberCell = targetCell.dataset.member.split(',');
+                Object.values(arrMembers).map((member) => {
+                    for (let i = 0; i<memberCell.length; i++){
+                        if (member.value === memberCell[i]){
+                            const itemMember = document.createElement('li');
+                            modalMember.append(itemMember)
+                            itemMember.append(`${member.label}`);
+                        }
                     }
-                }
-            })
+                })
+            } else {
+                modalMember.append('Event is empty!');
+            }
+            deleteButton.setAttribute('data-deleteId', eventId);
         }
 
         const del = document.getElementById('deleteEvent');
         del.addEventListener('click', () => {
             const delId = del.dataset.deleteid
-            console.log(delId)
-            deleteEventRequest(delId).then(()=>{
-                console.log("good", stateEvent)
-            }).catch((error)=>{
-                console.log("Error", error)
-            })
+            const result = confirm('Do you confirm the deletion of this event?');
+            if (result) {
+                deleteEventRequest(delId).then(() => {
+                    try {
+                        window.location.href = '/'
+                    } catch (e) {
+                        console.log("Error", e)
+                    }
+                })
+            }
         })
 
         function appendEventCell(){
@@ -192,7 +238,7 @@ if (window.location.pathname !== '/'){
                     timeCell.setAttribute('data-target', `#eventModal`)
 
                     timeCell.onclick = (event) => setEvent(event)
-                    timeCell.append('-')
+                    timeCell.append(' ')
                     dayRow.append(timeCell);
                 })
                 if (document.getElementById('calendarDays')){
@@ -210,18 +256,6 @@ if (window.location.pathname !== '/'){
             createWorkTime(listEvent);
         }
         createListEvent();
-
-        function createMemberList(){
-            const memberList = document.getElementById( 'eventMembers' ).getElementsByTagName( 'div' )[0];
-            Object.values(arrMembers).map((member)=>{
-                const memberItem = document.createElement('a');
-                memberItem.setAttribute('href', '#');
-                memberItem.setAttribute('class', 'dropdown-item');
-                memberItem.append(member.label)
-                memberList.appendChild(memberItem)
-            })
-        }
-        createMemberList();
 
         function createWorkTime(listEvent){
             Object.values(workTime).map((time, index) => {
@@ -259,20 +293,23 @@ if (window.location.pathname !== '/'){
             appendEventCell();
         }
         createDayEvent();
-
-        function CreateEventCell(data){
-            const eventsCell = document.getElementsByClassName('js-cell');
-            for(let k = 0; k<eventsCell.length; k++ ){
-                Object.values(data).map((item) => {
-                    if (eventsCell[k].dataset.day === item.day && eventsCell[k].dataset.time === item.time){
-                        eventsCell[k].innerHTML = '';
-                        eventsCell[k].setAttribute('data-member', item.member)
-                        eventsCell[k].setAttribute('data-id', item.id)
-                        eventsCell[k].append(item.name);
-                    }
-                })
-            }
-        }
         CreateEventCell(stateEvent);
     })
+
+    const instanceSort = new SelectPure(".sort_select", {
+        options: [{label: 'All members', value: '10'}, ...arrMembers],
+        multiple: false,
+        onChange: value => sortEventWithMember(value)
+    })
+
+    function sortEventWithMember(member){
+        let updateStateEvent = {}
+        const res = Object.entries(stateEvent).map((item, index) => {
+            if (item[1].member.filter((el) => el === member).length > 0){
+                updateStateEvent[item[0]] = item[1]
+            }
+        })
+        console.log(updateStateEvent)
+        CreateEventCell(updateStateEvent) //todo continued with sort member
+    }
 }
